requirements.txt
pyyaml
loguru
wandb
configmypy
numpy
torch
matplotlib
imageio
imageio-ffmpeg
pillow
torchgeometry
trimesh
pyglet==v1.5.28
libigl
kornia
plyfile
torchvision
opencv-python
